+++ analytics_toolkit/sql/metadata/__init__.py
@@ -1,4 +1,5 @@
 from .show_tables import show_tables
+from .show_queries import show_queries
 from .table_info import SqlTableInfo, table_info
 
-__all__ = ["SqlTableInfo", "show_tables", "table_info"]
+__all__ = ["SqlTableInfo", "show_queries", "show_tables", "table_info"]